commented some unused variables here and there
	modified:   Makefile.erebos
	modified:   loca_wrappers.f90
	modified:   par_solve_mumps.f90
	modified:   qc_sp_m.f90
	modified:   start_sparse_kit.f90

+++ par_solve_mumps.f90
@@ -57,7 +57,7 @@ MODULE par_solve_mumps
    INTEGER, PARAMETER :: CMPL_FLAG =  2
 
    ! memory parameters
-   REAL(KIND=8), PARAMETER :: memRelax = 5
+   INTEGER, PARAMETER :: memRelax = 5
 
 
    ! interface for real/complex matrix and rhs
